Added Tasks, Manager and Connected the files
There are new classes defined as Tasks which currently have only the two
properties - task type and target time. One field to be added is task
info that will serve as the data required to execute that task. You are
free to add random properties for now.

The manager is supposed to take care of the smallest task firsta and
update the global time accordingly.